+++ CHANGELOG.md
@@ -3,5 +3,11 @@
 ## 0.0.next
 - TBD
 
-## 0.0.1 (to be released on 2020.08)
-- TBD
+## 0.0.2 (to be released on 2020.08)
+- general waiting for entities
+  - via `new Wait(entity).for({toString: () => 'description', call: entity => someCodeOnEntityToWait() })`
+  - with some needed callables
+- Located class as draft of Lazy wrapper over locating both element and collection
+
+## 0.0.1 (released on 2020.08.15)
+- just a dummy package